Add field office phones for senators
Add a dump of senators' field office phone numbers.
Expose them in the API.
Preserve the original Contact struct, so that
this can be deployed without disrupting the existing
front-end.
Add a more exhaustive Phones field to Contact.
The intended use is that the UI will first show
the DC ("HQ") phone number.
If the user indicates that it is unavailable,
instead of giving up, we can instead ask them to
call the best field office (if it exists) followed
one or more of the other field offices.

Missing from this commit is a way to record
which field office they called in RegisterCall.
That will be kept separate, as it requires
a DB migration and other ugliness.