Suppress localStorage quota exceptions, fixes #126
In private browsing mode, Safari sets the storage quota to 0, which means localStorage throws an exception whenever you try to set anything. This wraps all storage setting code with error handling. The callback for setters also returns the error if there was one in case there’s a need for that later on.