add phone column to results
This way the client can report which phone
number the result is for, using the
"phone" param. The phone param is optional,
so that the backend can be deployed without
requiring upgraded clients.

Old data and old clients will result in empty
phone column entries.

This will run a migration on startup,
so if the production db is very large,
the first startup may be slow.