Use Choo to change url. Possible fix for #161
My brief reading of Choo documentation suggests we
should be using Choo's internal location model's reducer
when we want to change the url. Especially since we are
relying on it elsewhere to determine the params. See
https://github.com/yoshuawuyts/choo/blob/f84ec43fa58508cc20fe537d752a14901339f0cd/README.md#router
for more information.

With this change, selecting issues works for me
on IE 11 in Windows 7, fixing #161.